Revisi(membuka pdf, hapus file saat melalukan update, dan perubahan beberapa kondisi)

+++ app/Http/Controllers/ProfilPerusahaanContrl.php
@@ -18,28 +18,58 @@ class ProfilPerusahaanContrl extends Controller
 
     public function updateprofil(Request $request)
     {
-        $validatedData = $request->validate(
-            [
-                'nama_perusahaan' => 'required|max:30',
-                'alamat_perusahaan' => 'required|max:255',
-                'notlp_perusahaan' => 'required|max:30',
-                'deskripsi_perusahaan' => 'required',
-                'foto_perusahaan' => 'mimes:jpg,png|file|max:1524',
-                'surat_perusahaan' => 'mimes:pdf|file|max:1524',
-            ]
-        );
-        if ($request->foto_perusahaan != null) {
-            $validatedData['foto_perusahaan'] = $request->file('foto_perusahaan')->store('images-perusahaan');
-            Storage::delete(Auth::guard('perusahaan')->user()->foto_perusahaan);
-        } elseif ($request->surat_perusahaan != null) {
-            $validatedData['surat_perusahaan'] = $request->file('surat_perusahaan')->store('surat-perusahaan');
-            Storage::delete(Auth::guard('perusahaan')->user()->surat_perusahaan);
-        } elseif ($request->foto_perusahaan != null and $request->surat_perusahaan != null) {
+
+        if (Auth::guard('perusahaan')->user()->surat_perusahaan == null and Auth::guard('perusahaan')->user()->foto_perusahaan == null) {
+            $validatedData = $request->validate(
+                [
+                    'nama_perusahaan' => 'required|max:100',
+                    'alamat_perusahaan' => 'required|max:255',
+                    'notlp_perusahaan' => 'required|max:30',
+                    'deskripsi_perusahaan' => 'required',
+                    'foto_perusahaan' => 'required|mimes:jpg,png|file|max:1524',
+                    'surat_perusahaan' => 'required|mimes:pdf|file|max:1524',
+                ]
+            );
             $validatedData['foto_perusahaan'] = $request->file('foto_perusahaan')->store('images-perusahaan');
             $validatedData['surat_perusahaan'] = $request->file('surat_perusahaan')->store('surat-perusahaan');
-            Storage::delete(Auth::guard('perusahaan')->user()->foto_perusahaan);
-            Storage::delete(Auth::guard('perusahaan')->user()->surat_perusahaan);
+        } else {
+            $validatedData = $request->validate(
+                [
+                    'nama_perusahaan' => 'required|max:100',
+                    'alamat_perusahaan' => 'required|max:255',
+                    'notlp_perusahaan' => 'required|max:30',
+                    'deskripsi_perusahaan' => 'required',
+                    'foto_perusahaan' => 'mimes:jpg,png|file|max:1524',
+                    'surat_perusahaan' => 'mimes:pdf|file|max:1524',
+                ]
+            );
+            if ($request->foto_perusahaan != null) {
+                $validatedData['foto_perusahaan'] = $request->file('foto_perusahaan')->store('images-perusahaan');
+                Storage::delete(Auth::guard('perusahaan')->user()->foto_perusahaan);
+            } elseif ($request->surat_perusahaan != null) {
+                $validatedData['surat_perusahaan'] = $request->file('surat_perusahaan')->store('surat-perusahaan');
+                Storage::delete(Auth::guard('perusahaan')->user()->surat_perusahaan);
+            }
+            if ($request->foto_perusahaan != null and $request->surat_perusahaan != null) {
+                $validatedData['foto_perusahaan'] = $request->file('foto_perusahaan')->store('images-perusahaan');
+                $validatedData['surat_perusahaan'] = $request->file('surat_perusahaan')->store('surat-perusahaan');
+                Storage::delete(Auth::guard('perusahaan')->user()->foto_perusahaan);
+                Storage::delete(Auth::guard('perusahaan')->user()->surat_perusahaan);
+            }
         }
+
+        // if ($request->foto_perusahaan != null) {
+        //     $validatedData['foto_perusahaan'] = $request->file('foto_perusahaan')->store('images-perusahaan');
+        //     Storage::delete(Auth::guard('perusahaan')->user()->foto_perusahaan);
+        // } elseif ($request->surat_perusahaan != null) {
+        //     $validatedData['surat_perusahaan'] = $request->file('surat_perusahaan')->store('surat-perusahaan');
+        //     Storage::delete(Auth::guard('perusahaan')->user()->surat_perusahaan);
+        // } elseif ($request->foto_perusahaan != null and $request->surat_perusahaan != null) {
+        //     $validatedData['foto_perusahaan'] = $request->file('foto_perusahaan')->store('images-perusahaan');
+        //     $validatedData['surat_perusahaan'] = $request->file('surat_perusahaan')->store('surat-perusahaan');
+        //     Storage::delete(Auth::guard('perusahaan')->user()->foto_perusahaan);
+        //     Storage::delete(Auth::guard('perusahaan')->user()->surat_perusahaan);
+        // }
 
         // perusahaan::create($validatedData);
         perusahaan::where('id', Auth::guard('perusahaan')->user()->id)->update($validatedData);
@@ -92,5 +122,9 @@ class ProfilPerusahaanContrl extends Controller
     public function back()
     {
         return redirect()->back();
+    }
+    public function surat()
+    {
+        return view('perusahaan.dashboard.profil.surat');
     }
 }
